Move error parsing functions to shared helpers
