+++ src/applications/personalization/profile360/components/PaymentInformationEditModalError.jsx
@@ -1,15 +1,13 @@
 import React from 'react';
 import AlertBox from '@department-of-veterans-affairs/formation-react/AlertBox';
 
-// possible values for the `key` property on error messages we get from the server
-const ACCOUNT_FLAGGED_FOR_FRAUD_KEY = 'cnp.payment.flashes.on.record.message';
-const GENERIC_ERROR_KEY = 'cnp.payment.generic.error.message';
-const INVALID_ROUTING_NUMBER_KEY =
-  'payment.accountRoutingNumber.invalidCheckSum';
-const PAYMENT_RESTRICTIONS_PRESENT_KEY =
-  'payment.restriction.indicators.present';
-const ROUTING_NUMBER_FLAGGED_FOR_FRAUD_KEY =
-  'cnp.payment.routing.number.fraud.message';
+import {
+  hasFlaggedForFraudError,
+  hasInvalidAddressError,
+  hasInvalidHomePhoneNumberError,
+  hasInvalidRoutingNumberError,
+  hasInvalidWorkPhoneNumberError,
+} from '../util';
 
 function FlaggedAccount() {
   return (
@@ -88,77 +86,6 @@ function UpdatePhoneNumberError({ closeModal, phoneNumberType = 'home' }) {
       and fill in this required information.
     </p>
   );
-}
-
-function hasErrorMessageKey(errors, errorKey) {
-  return errors.some(err =>
-    err.meta.messages.some(message => message.key === errorKey),
-  );
-}
-
-function hasErrorMessageText(errors, errorText) {
-  return errors.some(err =>
-    err.meta.messages.some(message =>
-      message.text.toLowerCase().includes(errorText.toLowerCase()),
-    ),
-  );
-}
-
-function hasFlaggedForFraudError(errors) {
-  return (
-    hasErrorMessageKey(errors, ACCOUNT_FLAGGED_FOR_FRAUD_KEY) ||
-    hasErrorMessageKey(errors, PAYMENT_RESTRICTIONS_PRESENT_KEY) ||
-    hasErrorMessageKey(errors, ROUTING_NUMBER_FLAGGED_FOR_FRAUD_KEY)
-  );
-}
-
-function hasInvalidRoutingNumberError(errors) {
-  let result = false;
-  if (hasErrorMessageKey(errors, INVALID_ROUTING_NUMBER_KEY)) {
-    result = true;
-  }
-  if (
-    hasErrorMessageKey(errors, GENERIC_ERROR_KEY) &&
-    hasErrorMessageText(errors, 'Invalid Routing Number')
-  ) {
-    result = true;
-  }
-  return result;
-}
-
-function hasInvalidAddressError(errors) {
-  let result = false;
-  if (
-    hasErrorMessageKey(errors, GENERIC_ERROR_KEY) &&
-    hasErrorMessageText(errors, 'address update')
-  ) {
-    result = true;
-  }
-  return result;
-}
-
-function hasInvalidHomePhoneNumberError(errors) {
-  let result = false;
-  if (
-    hasErrorMessageKey(errors, GENERIC_ERROR_KEY) &&
-    (hasErrorMessageText(errors, 'night phone number') ||
-      hasErrorMessageText(errors, 'night area number'))
-  ) {
-    result = true;
-  }
-  return result;
-}
-
-function hasInvalidWorkPhoneNumberError(errors) {
-  let result = false;
-  if (
-    hasErrorMessageKey(errors, GENERIC_ERROR_KEY) &&
-    (hasErrorMessageText(errors, 'day phone number') ||
-      hasErrorMessageText(errors, 'day area number'))
-  ) {
-    result = true;
-  }
-  return result;
 }
 
 export default function PaymentInformationEditModalError({
